+++ src/main/common/log.c
@@ -46,6 +46,10 @@
 #include "msp/msp_serial.h"
 #include "msp/msp_protocol.h"
 
+#if defined(WASM_BUILD)
+#include <emscripten/console.h>
+#endif
+
 #if defined(USE_LOG)
 
 #define LOG_PREFIX                  "[%6d.%03d] "
@@ -131,6 +135,11 @@ static void logPrint(const char *buf, size_t size)
         fputc(buf[ii], stdout);
     }
 #endif
+
+#if defined(WASM_BUILD)
+    emscripten_console_log(buf);
+#endif
+
     SD(printf("%s\n", buf));
     if (logPort) {
         // Send data via UART (if configured & connected - a safeguard against zombie VCP)
