+++ src/main/config/parameter_group.c
@@ -24,15 +24,30 @@
 #include "parameter_group.h"
 #include "common/maths.h"
 
-/* WebAssembly / Emscripten fallback symbols
-   When weak symbols are used and not provided by linker, define empty arrays
-   This is needed for WASM builds where linker scripts may not work correctly
+/* WebAssembly / Emscripten: Dynamic registry array
+   For WASM, we use an array-based registry instead of linker sections.
+   Registries are added at runtime via constructor functions.
 */
-#ifdef __EMSCRIPTEN__
-const pgRegistry_t __pg_registry_start[1] __attribute__((weak)) = {0};
-const pgRegistry_t __pg_registry_end[1] __attribute__((weak)) = {0};
-const uint8_t __pg_resetdata_start[1] __attribute__((weak)) = {0};
-const uint8_t __pg_resetdata_end[1] __attribute__((weak)) = {0};
+#if defined(WASM_BUILD)
+    #define PG_REGISTRY_MAX_ENTRIES 256
+    
+    static pgRegistry_t __pg_registry_data[PG_REGISTRY_MAX_ENTRIES];
+    static size_t __pg_registry_count = 0;
+    
+    const pgRegistry_t* __pg_registry_start = __pg_registry_data;
+    const pgRegistry_t* __pg_registry_end = __pg_registry_data;  // Updated after all registrations
+    
+    void pgRegistryAdd(const pgRegistry_t *reg)
+    {
+        if (__pg_registry_count >= PG_REGISTRY_MAX_ENTRIES) {
+            // Error: too many registries. Increase PG_REGISTRY_MAX_ENTRIES
+            return;
+        }
+        __pg_registry_data[__pg_registry_count++] = *reg;
+        // Update end pointer to reflect new count
+        __pg_registry_end = &__pg_registry_data[__pg_registry_count];
+    }
+
 #endif
 
 const pgRegistry_t* pgFind(pgn_t pgn)
@@ -61,7 +76,12 @@ static void pgResetInstance(const pgRegistry_t *reg, uint8_t *base)
     const uint16_t regSize = pgSize(reg);
 
     memset(base, 0, regSize);
+    
+#if defined(WASM_BUILD)
+    if (reg->reset.ptr) {
+#else
     if (reg->reset.ptr >= (void*)__pg_resetdata_start && reg->reset.ptr < (void*)__pg_resetdata_end) {
+#endif
         // pointer points to resetdata section, to it is data template
         memcpy(base, reg->reset.ptr, regSize);
     } else if (reg->reset.fn) {
